requirements.txt
pillow>=5.3.0